requirements.txt
# Aero-Gauge Requirements
# Python 3.10+

# Core ML/DL
torch>=2.6.0
torchvision>=0.21.0
torchaudio>=2.6.0

# Computer Vision
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Data Science
numpy>=2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Web App
streamlit>=1.29.0

# Progress bars
tqdm>=4.66.0

# Optional: For Kaggle dataset download
# kaggle>=1.5.16

# Optional: Advanced models (EfficientNet)
# timm>=0.9.12
